Finished running training for age and motion QC score.
