+++ bin/dcan/run_tensorboard.sh
@@ -1,0 +1,1 @@
+/home/miran045/reine097/.local/bin/tensorboard $1
