+++ src/dcan/dsets/age.py
@@ -1,0 +1,67 @@
+import copy
+import random
+
+import torch
+import torch.cuda
+from torch.utils.data import Dataset
+
+from dcan.dsets.dsets import getCandidateInfoList, getMriRawCandidate
+from util.logconf import logging
+
+log = logging.getLogger(__name__)
+# log.setLevel(logging.WARN)
+# log.setLevel(logging.INFO)
+log.setLevel(logging.DEBUG)
+
+class MRIAgeDataset(Dataset):
+    def __init__(self,
+                 val_stride=0,
+                 isValSet_bool=None,
+                 series_uid=None,
+                 sortby_str='random',
+            ):
+        self.candidateInfo_list = copy.copy(getCandidateInfoList())
+
+        if series_uid:
+            self.candidateInfo_list = [
+                x for x in self.candidateInfo_list if x.series_uid == series_uid
+            ]
+
+        if isValSet_bool:
+            assert val_stride > 0, val_stride
+            self.candidateInfo_list = self.candidateInfo_list[::val_stride]
+            assert self.candidateInfo_list
+        elif val_stride > 0:
+            del self.candidateInfo_list[::val_stride]
+            assert self.candidateInfo_list
+
+        if sortby_str == 'random':
+            random.shuffle(self.candidateInfo_list)
+        elif sortby_str == 'series_uid':
+            self.candidateInfo_list.sort(key=lambda x: (x.series_uid, x.center_xyz))
+        elif sortby_str == 'label_and_size':
+            pass
+        else:
+            raise Exception("Unknown sort: " + repr(sortby_str))
+
+        log.info("{!r}: {} {} samples".format(
+            self,
+            len(self.candidateInfo_list),
+            "validation" if isValSet_bool else "training",
+        ))
+
+    def __len__(self):
+        return len(self.candidateInfo_list)
+
+    def __getitem__(self, ndx):
+        candidateInfo_tup = self.candidateInfo_list[ndx]
+
+        candidate_a = getMriRawCandidate(
+            candidateInfo_tup.series_uid,
+        )
+        candidate_t = torch.from_numpy(candidate_a).to(torch.float32)
+        candidate_t = candidate_t.unsqueeze(0)
+
+        age_t = torch.tensor(candidateInfo_tup.age_int, dtype=torch.double)
+
+        return candidate_t, age_t, candidateInfo_tup.series_uid
