+++ src/dcan/model/alex_net.sh
@@ -1,8 +1,9 @@
 #!/bin/sh
 
-#SBATCH --job-name=training-alex-net # job name
-#SBATCH --mem=90g        # memory per cpu-core (what is the default?)
-#SBATCH --time=16:00:00          # total run time limit (HH:MM:SS)
+#SBATCH --job-name=alex-net # job name
+
+#SBATCH --mem=90G        # memory per cpu-core (what is the default?)
+#SBATCH --time=8:00:00          # total run time limit (HH:MM:SS)
 #SBATCH -p v100
 #SBATCH --gres=gpu:v100:1
 #SBATCH --ntasks=6               # total number of tasks across all nodes
@@ -10,13 +11,12 @@
 #SBATCH --mail-type=begin        # send 7mail when job begins
 #SBATCH --mail-type=end          # send email when job ends
 #SBATCH --mail-user=reine097@umn.edu
-#SBATCH -e alex-net-training-%j.err
-#SBATCH -o alex-net-training-%j.out
+#SBATCH -e alex-net-%j.err
+#SBATCH -o alex-net-%j.out
 
-#SBATCH -A miran045
+#SBATCH -A rando149
 
 cd /home/miran045/reine097/projects/AlexNet_Abrol2021 || exit
 export PYTHONPATH=PYTHONPATH:"/home/miran045/reine097/projects/AlexNet_Abrol2021/src"
 /home/miran045/reine097/projects/AlexNet_Abrol2021/venv/bin/python \
-  /home/miran045/reine097/projects/AlexNet_Abrol2021/src/dcan/training.py --num-workers=4 --batch-size=8 \
-  --tb-prefix="age_regression" --epochs=100 "dcan"
+  /home/miran045/reine097/projects/AlexNet_Abrol2021/src/dcan/model/alex_net.py
